==> Week 1/03-similarity-langchain.ipynb <==
# Load data

import glob
import os

!pip install langchain

!pip install langchain-community

from langchain.document_loaders import DirectoryLoader
from langchain.document_loaders.text import TextLoader
loader = DirectoryLoader("un/TXT/Session 77 - 2022/", glob="**/*.txt", loader_cls=TextLoader)
data = loader.load()

data[0:5]

from langchain.text_splitter import RecursiveCharacterTextSplitter

text_splitter = RecursiveCharacterTextSplitter(chunk_size = 500, chunk_overlap = 50)
all_splits = text_splitter.split_documents(data)

all_splits[0:10]

# this used to work
# from langchain_huggingface import HuggingFaceEmbeddings
# this is deprecated
from langchain.embeddings import HuggingFaceEmbeddings, SentenceTransformerEmbeddings
# this will work (now and) in the future
# from langchain_huggingface import HuggingFaceEmbeddings
from langchain.vectorstores.sklearn import SKLearnVectorStore

retriever = SKLearnVectorStore.from_documents(all_splits, HuggingFaceEmbeddings(model_name="all-MiniLM-L6-v2"))\
                              .as_retriever(search_kwargs={"k": 100})

question = "The climate crisis is worse for poorer countries"
docs = retriever.invoke(question)
len(docs)

docs[0:10]

import pandas as pd
pd.set_option('display.max_colwidth', 0)
df = pd.DataFrame([{"source": d.metadata["source"], "text": d.page_content} for d in docs])
df

from langchain_community.cross_encoders import HuggingFaceCrossEncoder
from langchain.retrievers.document_compressors import CrossEncoderReranker
from langchain.retrievers import ContextualCompressionRetriever


cross_encoder = HuggingFaceCrossEncoder(model_name="mixedbread-ai/mxbai-rerank-large-v1")
compressor = CrossEncoderReranker(model=cross_encoder, top_n=20)
compression_retriever = ContextualCompressionRetriever(
    base_compressor=compressor, base_retriever=retriever
)

compressed_docs = compression_retriever.invoke(question)

pd.DataFrame([{"source": d.metadata["source"], "text": d.page_content} for d in compressed_docs])
