==> Week 1/04-txtai.ipynb <==
# Load data (from previous notebook)

sentences = open("sentences.txt").read().split("@@@")

len(sentences)

# currently does not work without staticvectors
!pip install txtai staticvectors

from txtai.embeddings import Embeddings

embeddings = Embeddings({"path": "sentence-transformers/multi-qa-MiniLM-L6-cos-v1", "content": True})

embeddings.index(sentences)

query = "The climate crisis is worse for poorer countries"
res = embeddings.search(query, 100)
res[0:10]

import pandas as pd
pd.set_option('display.max_colwidth', 0)
df = pd.DataFrame(res)
df

**Identical** results to our first notebook!

from txtai.pipeline import Similarity
#similarity = Similarity('mixedbread-ai/mxbai-rerank-large-v1')
similarity = Similarity('valhalla/distilbart-mnli-12-3')

pd.DataFrame([(res[x]["id"], res[x]["text"], res[x]["score"], score) 
 for x, score in similarity(query, [t["text"] for t in res])], 
            columns=["id", "text", "bi-score", "cross-score"])
